Do the same for the sky and move these into the render loop.
